+++ CHANGELOG.md
@@ -94,6 +94,7 @@ Changelog
 - Redundant checksum files are no longer published: checksums of `.asc` signature files ([gradle/gradle#20232](https://github.com/gradle/gradle/issues/20232)) and the `sha256`/`sha512`. If you rely on these, let me know.
 - Remove `binding-contributions-as-containers` flag, only `binding-contributions-as-containers` is supported going forward.
 - **[gradle]** Build against Gradle `9.6.0`.
+- Test Kotlin `2.4.20-Beta1`.
 
 ### Contributors